+++ R/utils_link_inverse.R
@@ -1,35 +1,4 @@
 #' @importFrom stats family
-get_link_inverse <- function(fun, model) {
-
-  # handle glmmTMB models
-  if (inherits(model, "glmmTMB")) {
-    ff <- stats::family(model)
-
-    if ("linkinv" %in% names(ff))
-      return(ff$linkinv)
-    else
-      return(match.fun("exp"))
-  }
-
-  # do we have glm? if so, get link family. make exceptions
-  # for specific models that don't have family function
-  if (any(fun %in% c("lme", "plm", "gls", "lm", "truncreg", "zeroinfl", "hurdle", "coxph"))) {
-    il <- NULL
-  } else if (fun %in% c("betareg")) {
-    il <- model$link$mean$linkinv
-  } else if (fun %in% c("lrm", "polr")) {
-    # "lrm"-object from pkg "rms" have no family method
-    # so we construct a logistic-regression-family-object
-    il <- stats::binomial(link = "logit")$linkinv
-  } else {
-    # get family info
-    il <- stats::family(model)$linkinv
-  }
-
-  il
-}
-
-
 get_link_fun <- function(model) {
   # get model family
   ff <- stats::family(model)
@@ -41,4 +10,20 @@ get_link_fun <- function(model) {
   if ("link" %in% names(ff)) return(match.fun(ff$link))
 
   NULL
+}
+
+
+#' @importFrom stats terms weights
+get_model_weights <- function(model) {
+  w <- NULL
+  tryCatch(
+    {
+      w <- stats::weights(model)
+    },
+    error = function(x) { NULL },
+    warning = function(x) { NULL },
+    finally = function(x) { NULL }
+  )
+
+  w
 }