+++ R/utils_model_family.R
@@ -1,8 +1,13 @@
-#' @importFrom sjmisc str_contains
+#' @importFrom sjmisc str_contains is_empty
 #' @importFrom stats family
 get_glm_family <- function(fit) {
   # get model class
   mc <- get_predict_function(fit)
+
+  # for gam-components from gamm4, add class attributes, so family
+  # function works correctly
+  if (any(mc == "gam") && !any(c("glm", "lm") %in% mc))
+    class(fit) <- c(class(fit), "glm", "lm")
 
   # do we have glm? if so, get link family. make exceptions
   # for specific models that don't have family function
@@ -30,7 +35,7 @@ get_glm_family <- function(fit) {
   } else {
     # "lrm"-object from pkg "rms" have no family method
     # so we construct a logistic-regression-family-object
-    if (mc %in% c("lrm", "polr", "logistf"))
+    if (mc %in% c("lrm", "polr", "logistf", "clm", "multinom"))
       faminfo <- stats::binomial(link = "logit")
     else
       # get family info
@@ -42,19 +47,30 @@ get_glm_family <- function(fit) {
   }
 
   # create logical for family
-  binom_fam <- fitfam %in% c("binomial", "quasibinomial", "binomialff")
-  poisson_fam <- fitfam %in% c("poisson", "quasipoisson")
-  neg_bin_fam <- sjmisc::str_contains(fitfam, "negative binomial", ignore.case = T)
+  binom_fam <- any(
+    fitfam %in% c("binomial", "quasibinomial", "binomialff") |
+    sjmisc::str_contains(fitfam, "binomial", ignore.case = TRUE)
+  )
 
-  return(
-    list(
-      is_bin = binom_fam,
-      is_pois = poisson_fam | neg_bin_fam,
-      is_negbin = neg_bin_fam,
-      is_logit = logit_link,
-      link.fun = link.fun,
-      family = fitfam
-    )
+  poisson_fam <- any(
+    fitfam %in% c("poisson", "quasipoisson") |
+    sjmisc::str_contains(fitfam, "poisson", ignore.case = TRUE)
+  )
+
+  neg_bin_fam <- any(
+    sjmisc::str_contains(fitfam, "negative binomial", ignore.case = T) |
+    sjmisc::str_contains(fitfam, "nbinom", ignore.case = TRUE) |
+    sjmisc::str_contains(fitfam, "neg_binomial", ignore.case = TRUE)
+  )
+
+
+  list(
+    is_bin = binom_fam & !neg_bin_fam,
+    is_pois = poisson_fam | neg_bin_fam,
+    is_negbin = neg_bin_fam,
+    is_logit = logit_link,
+    link.fun = link.fun,
+    family = fitfam
   )
 }